Refactored BaseStateNotifierTests
